Separated test class due to lines exceeded 1000
